+++ Sources/MapboxMaps/Foundation/MapViewDependencyProvider.swift
@@ -28,7 +28,7 @@ internal final class MapViewDependencyProvider: MapViewDependencyProviderProtoco
     }
 
     func makePinchGestureHandler(view: UIView,
-                                 mapboxMap: MapboxMapProtocol) -> GestureHandler {
+                                 mapboxMap: MapboxMapProtocol) -> PinchGestureHandlerProtocol {
         let gestureRecognizer = UIPinchGestureRecognizer()
         view.addGestureRecognizer(gestureRecognizer)
         return PinchGestureHandler(