+++ Tests/MapboxMapsTests/Gestures/GestureHandlers/Mocks/MockPinchGestureHandler.swift
@@ -1,0 +1,5 @@
+@testable import MapboxMaps
+
+final class MockPinchGestureHandler: GestureHandler, PinchGestureHandlerProtocol {
+    var rotateEnabled: Bool = true
+}